refactor(canvas): unify Canvas via Backing enum, enabling real drawing
Previously, skia-rs-canvas had two parallel canvas types. Canvas was
the name every external consumer used (re-exported from skia-rs-safe
and the umbrella crate), but every draw method was a TODO stub and it
owned no pixel buffer. RasterCanvas was the real renderer but had to
be obtained via Surface::raster_canvas. Surface::canvas returned the
stub, unusable. Picture::playback took &mut Canvas, so every playback
draw was silently dropped. That left the crate's advertised Canvas API
shipping no pixels and pictures acting as write-only sinks.

Unify the two by making Canvas<'a> carry a Backing<'a> enum:

  Backing::Raster(&'a mut PixelBuffer)
      Draws dispatch to a freshly configured Rasterizer against the
      borrowed buffer using the current matrix and clip.
  Backing::Recording(&'a mut Vec<DrawCommand>)
      Draws (and matrix / clip / save / restore calls) push DrawCommand
      variants for later playback, giving recording a first-class
      backing without the old parallel RecordingCanvas pointer cast.
  Backing::Null
      Draws are discarded. Useful for quick_reject probes, CPU-only
      benchmarks, and deferred / culled render paths. Canvas::new
      (width, height) now returns a Null canvas so existing bench
      callers keep working.

Every draw method (clear, draw_color, draw_point, draw_points,
draw_line, draw_rect, draw_oval, draw_circle, draw_arc,
draw_round_rect, draw_path, draw_picture, plus draw_image* /
draw_region / draw_vertices / draw_text_blob) matches on
self.backing and routes to the correct code path. Clip and matrix
stack methods remain backing-agnostic and, on the recording backing,
record their commands for replay.

The with_rasterizer helper builds a Rasterizer pre-configured with
the current matrix and simple rect clip so the Raster arms stay
short. Arc and round-rect path construction are lifted into private
helpers so the pixel output matches the old RasterCanvas byte-for-
byte.

Surface::canvas now takes &mut self and returns the real raster
canvas; raster_canvas remains as a back-compat shim and RasterCanvas
is kept as a deprecated type alias to Canvas. Callers in skia-rs-svg
and skia-rs-safe migrated to Canvas directly; the deprecated alias
and shim are only for third-party source compatibility.

Drive-by fixes required to keep the workspace green:
- skia-rs-safe's text and codec features now forward to
  skia-rs-canvas so the text_rendering example builds when either
  feature is enabled.
- The bench `canvas` benchmark can no longer return the Canvas from
  iter_batched (it now borrows from the Surface), so the closure
  drops it in place.
- skia-rs/src/lib.rs's test_version hard-coded 0.2.2; switched to
  env!(CARGO_PKG_VERSION) so it tracks the workspace version.

Unblocks GAP-C1..C-7, C-9 from the skia-rs-canvas audit. GAP-C10
(save_layer composition), GAP-C8 (clip Difference), and GAP-C11
(flush semantics) are tracked separately by P5-4, P5-2, and P5-5.

New tests pin the end-to-end behavior: test_picture_playback_draws
_pixels records a clear + filled rect and replays to a raster
surface, test_recording_backing_captures_commands verifies the
recording path, test_null_canvas_discards_draws pins the null
contract, and test_surface_canvas_returns_functional_raster pins the
Surface::canvas fix.
